client/test: Move handlers from multiledger.go -> handler.go
